Adding relevance helper to reduce code duplication
