+++ src/__init__.py
@@ -28,6 +28,7 @@ _EXPOSED = {
     "pubmed_fetcher",    # needs BioPython – container provides it
     "prompt_library",    # no heavy deps
     "scoring_guidelines",# no heavy deps
+    "relevance_helper",  # shared helpers used by both relevance_triton and relevance_chtc
     "relevance_triton",  # Triton remote inference (with CHTC fallback)
     "relevance_chtc",    # vLLM / torch – GPU container only (standalone for CHTC jobs)
 }
